[feat/refactor] pubsub now sends [env, proto, ts]
The major change in this CL is that pubsub now publishes messages in the format:

[env, proto, ts]

where previously we did not append timestamps. The timestamp ts is created
when the message is published. This is used for an important
feature we are adding: filtering 'cached' messages. When a new component
subscribes to the pubsubcache, it detects and sends out all 'cached'
messages. However, these messages are published to everyone! This means that
on new subscriptions other components get spammed with old data.

To avoid this, we now have updated the Subscriber to track the latest timestamp
and filter out old messages it receives. In this way, old components do not get
spammed.

The main associated changes are:
- The pubsubcache now stores (proto, ts) cached tuples, where it used to store
just the protos. This makes significant changes throughout, wherever the cache
is accessed.
- In test_pubsub_logic, we now explicitly test that old components *do not*
receive cached messages (see test_pubsubcache_interaction).
- We add some helper functions: extract_ts(), get_cache_item(),
  get_cache_items(), get_ts().
- We generalize creating the message packet into create_message_packet().