[feat] update scheduling logic
Previously, we had the logic set up such that the component in control
lost control *whenever* a new problem was logged. Now, we do the
following:
- If the component in control solved EP_NONE (i.e. was generic), we
release its control when a problem is logged. In other words, whenever
we go from no problems to a problem, a component loses control.
- If the the component in control solves EP_XXX and that problem is
removed, we release its control.