[refactor] initial refactor of SXM translator
This hasn't been touched since Erwan pushed something a while back. Now
that we are wanting to integrate it, we decided to adapt it to use the
ConfigTranslator (and in principle support spectroscopies). This CL is a
first-pass at doing so.

We now follow the more standard file structure, with:
- an actions.py for the ActionHandler, and an associated actions.toml.
- a params.py for the ParameterHandler, and an associated params.toml.
- the main translator.py.

We removed scan.py, as it's main method was pulled into translator.py.

Note a number of things that are not finished:
- The readers for scans and spectroscopies are still empty, as we need
to play with the scifireaders reader and understand the save format
better.
- We need find out how to check for SS_SPEC state.
- We need to see if there are autosave/repeat options for scans (as
there are for specs).