[feat] updats to sxm reader
Here, we do a number of main changes:
1. Add a reader file that (we believe) properly reads SXM spec and
scans. We are still missing UTs!
2. Now import a local sxm file, which contains the DDEClient
interfacing. This is a modified version of the SXMRemote that comes with
the SXM software. Note that this is *not* yet pushed globally, as we are
waiting for confirmation we have the rights to push this. For now, it
will stay as a stash locally.
3. Update params.py to use the new sxm, and updates to the toml
accordingly.

Note also that our spec/scan reading sets attributes (probe position,
rotation angle) directly. But we suggest using config_translator's
correct_scan()/correct_spec() instead (to ensure consistent coordinate
system things). We will remove these calls in a follow-up CL.

TODO:
- Push sxm.py (stashed locally).
- UTs for reader.
- Remove direct attribute access for spec/scan.